+++ src/HashSearch/Hash/hashMethodCurtailing.java
@@ -1,7 +1,7 @@
 package HashSearch.Hash;
 
 public class hashMethodCurtailing implements Hash{
-    public int getHash(long value, int countAddress) {
+    public int getHash(int value, int countAddress) {
         int lengthAddress = Integer.toString(countAddress - 1).length(); // порядок числа адресов
         int lengthValue = Long.toString(value).length();
         int result = 0;
