+++ src/HashSearch/Hash/hashMethodDividing.java
@@ -1,7 +1,7 @@
 package HashSearch.Hash;
 
 public class hashMethodDividing implements Hash{
-    public int getHash(long value, int countAddress) {
+    public int getHash(int value, int countAddress) {
         return (int) ((value % divisor) * ((double)countAddress / divisor));
     }
 }
