+++ src/HashSearch/Hash/hashMethodMultipl.java
@@ -1,7 +1,7 @@
 package HashSearch.Hash;
 
 public class hashMethodMultipl implements Hash{
-    public int getHash(long value, int countAddress) {
+    public int getHash(int value, int countAddress) {
         double i = value * multiplier;
         i = i - (int) i;    // Дробная часть числа
         return (int) (countAddress * i);
